[chore] Move components folder to src folder
