+++ src/app/page.tsx
@@ -1,5 +1,5 @@
 import RootLayout from "./layout";
-import Header from "./components/Header/Header";
+import Header from "../components/Header/Header";
 
 export default function Home() {
 	return (
